+++ src/blocks/blockfurnace.h
@@ -41,7 +41,7 @@ class BlockFurnace : public BlockBasic
 public:
   inline bool affectedBlock(int block) const { return block == BLOCK_FURNACE || block == BLOCK_BURNING_FURNACE; }
 
-  bool onPlace(User* user, int16_t newblock, int32_t x, int8_t y, int32_t z, int map, int8_t direction);
-  bool onInteract(User* user, int32_t x, int8_t y, int32_t z, int map);
-  bool onBroken(User* user, int8_t status, int32_t x, int8_t y, int32_t z, int map,  int8_t direction);
+  bool onPlace(User* user, int16_t newblock, int32_t x, int16_t y, int32_t z, int map, int8_t direction);
+  bool onInteract(User* user, int32_t x, int16_t y, int32_t z, int map);
+  bool onBroken(User* user, int8_t status, int32_t x, int16_t y, int32_t z, int map,  int8_t direction);
 };
