+++ src/blocks/ice.cpp
@@ -31,14 +31,14 @@
 
 #include "ice.h"
 
-bool BlockIce::onBroken(User* user, int8_t status, int32_t x, int8_t y, int32_t z, int map, int8_t direction)
+bool BlockIce::onBroken(User* user, int8_t status, int32_t x, int16_t y, int32_t z, int map, int8_t direction)
 {
   ServerInstance->map(map)->setBlock(x, y, z, BLOCK_WATER, direction);
   ServerInstance->map(map)->sendBlockChange(x, y, z, BLOCK_WATER, direction);
   return true;
 }
 
-bool BlockIce::onPlace(User* user, int16_t newblock, int32_t x, int8_t y, int32_t z, int map, int8_t direction)
+bool BlockIce::onPlace(User* user, int16_t newblock, int32_t x, int16_t y, int32_t z, int map, int8_t direction)
 {
   uint8_t oldblock;
   uint8_t oldmeta;
